terraform: Compare locks and provider requirements
When building a context, we read the dependency locks and ensure that
the provider requirements from the configuration can be satisfied.
If the configured requirements change such that the locks need to be
updated, we explain this and recommend running "terraform init".

This check is disabled if either of the provider developer tools are in
place. This includes unmanaged providers and the provider installation
`dev_overrides` block.