Update EvalValidateResource
Return diagnostics rather than an error, since the EvalNodes have been
removed and the caller has been updated.

IgnoreWarnings was no longer used.

The validation for ignore_changes was too broad, and makes it difficult
to ignore changes in an attribute that the user does not want to set.
While the goal of ignore_changes is to prevent changes in the
configuration, we don't intend to break the use-case of ignoring drift
from the provider. Since we cannot easily narrow the validation to only
detect computed attributes at the moment, we can drop this error
altogether for now.